+++ test/nextgen/reference/compare.cpp
@@ -15,6 +15,7 @@
 
 #include "test/common/assert.hpp"
 #include "test/common/data_type.hpp"
+#include "test/common/float16.hpp"
 #include "test/common/int4.hpp"
 #include "test/common/memory.hpp"
 #include "test/common/span.hpp"
@@ -98,6 +99,9 @@ CompareFn make_compare_plain_2d(DataType dtype) {
     switch (dtype) {
         case DataType::FP32:
             return compare_plain_2d<float>;
+
+        case DataType::FP16:
+            return compare_plain_2d<Float16>;
 
         case DataType::I32:
             return compare_plain_2d<int32_t>;